LayoutTests/js/add-recovery-expected.txt

This tests that speculation recovery of destructive additions on unboxed integers works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS destructiveAddForBoxedInteger({x:1}, 2, 4) is 9
PASS destructiveAddForBoxedInteger({x:2147483647}, 2, 4) is 2147483655
PASS destructiveAddForBoxedInteger({x:2}, 2147483647, 4) is 4294967300
PASS destructiveAddForBoxedInteger({x:2147483647}, 2147483647, 4) is 6442450945
PASS destructiveAddForBoxedInteger({x:1}, 2, 2147483647) is 2147483652
PASS successfullyParsed is true

TEST COMPLETE
